rearrange components in stream testing panel
