+++ airbyte-webapp/src/components/StreamTestingPanel/StreamTester.module.scss
@@ -1,0 +1,32 @@
+@use "scss/variables";
+@use "scss/colors";
+
+.container {
+  flex: 1;
+  display: flex;
+  flex-direction: column;
+}
+
+.resizablePanelsContainer {
+  flex: 1;
+  min-height: 0;
+
+  // required to hide the connector logs splitter underneath the resizable panel overlay
+  z-index: 0;
+}
+
+.testButtonText {
+  color: white;
+}
+
+// .error {
+//   align-self: center;
+//   width: 70%;
+//   background-color: colors.$red-50;
+//   color: colors.$red;
+//   margin-top: variables.$spacing-xl;
+//   border-radius: variables.$border-radius-sm;
+//   padding: variables.$spacing-xl;
+//   display: flex;
+//   align-items: center;
+// }
